Use a release build, compile wasm-ld, and don't build tools
Debug builds take up enormous amounts of memory when linking due to the
generation of DWARF information. We might also need wasm-ld to link
object files together. Lastly, we should try not to build unnecessary
tools and binaries, so we disable building tools as a first step in that
direction.