Lunatic and wasi (#8)
* Use llvm-js as a dependency via a tarball

npm hates git for whatever reason.

* Add declarations for every Lunatic API in existence

There are some TODOs related to enums, but this should otherwise work
very well. Higher-level bindings are of course necessary.

* Add declarations for WASI

Enums and structs should be added in the future, but this should provide
all of the necessary imports for using the WASI syscalls. Higher-level
bindings need to be implemented, of course.

* Add a test file for testing Lunatic and WASI compilation

The code is mostly bogus, but it should theoretically compile.

* Don't erroneously report a diagnostic after pushing a scope element

Someone forgot a `return;`. :^)

* lots of fixes

* enums work

* something is wrong

* fixed the bc generation

* test files, build script

* Small fixes

---------

Co-authored-by: CountBleck <Mr.YouKnowWhoIAm@protonmail.com>